Change type of operands

+++ src/machine/behavior/is/float/fune.rs
@@ -3,9 +3,9 @@ use machine::state::sr::R;
 
 pub fn fune(state: &mut State, x: u8, y: u8, z: u8) {
     // Load operands
-    let op1: u64 = state.gpr[y].into();
-    let op2: u64 = state.gpr[z].into();
-    let eps: u64 = state.sr[R::E].into();
+    let op1: f64 = state.gpr[y].into();
+    let op2: f64 = state.gpr[z].into();
+    let eps: f64 = state.sr[R::E].into();
 
     // Execute
     let res = ((op1 as i64).wrapping_sub(op2 as i64)).abs();
